서비스 엔드포인트 관련 로직 추가

+++ src/main/java/kr/co/strato/portal/workload/controller/StatefulSetDetailController.java
@@ -19,7 +19,7 @@ public class StatefulSetDetailController {
 
     @DeleteMapping("api/v1/statefulset-detail/statefulsets/{id}")
     @ResponseStatus(HttpStatus.OK)
-    public ResponseWrapper<Boolean> deleteStatefulSet(@PathVariable Integer id){
+    public ResponseWrapper<Boolean> deleteStatefulSet(@PathVariable Long id){
         boolean isDeleted = statefulSetDetailService.deleteStatefulSet(id);
 
         return new ResponseWrapper<>(isDeleted);
